+++ index.html
@@ -230,7 +230,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src="js/main.js">
+  <script src="js/main.js" type="module">
   </script>
 </body>
 </html>
